Added fixes by mentor`s comments
