+++ project/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.5.0",
     "axios": "^0.21.1",
+    "dayjs": "1.10.6",
     "history": "^4.10.1",
     "leaflet": "1.7.1",
     "prop-types": "^15.7.2",
