+++ project/src/components/app/app.jsx
@@ -1,17 +1,17 @@
 import React, { useEffect } from 'react';
-import MainPage from '../../routes/main-page';
+import MainPage from '../../routes/main-page/main-page';
 import { Router as BrowserRouter, Switch, Route } from 'react-router-dom';
 import { AppRoute } from '../../constants';
-import LoginScreen from '../../routes/login';
-import Favorites from '../../routes/favorites';
+import LoginScreen from '../../routes/login/login';
+import Favorites from '../../routes/favorites/favorites';
 import Room from '../../routes/room/room';
-import NotFoundScreen from '../../routes/not-found';
-import LoadingScreen from '../loading-screen';
-import { isCheckedAuth } from '../../utils';
+import NotFoundScreen from '../../routes/not-found/not-found';
+import LoadingScreen from '../loading-screen/loading-screen';
+import { isCheckedAuth } from '../../utils/utils';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import PrivateRouteToFavorite from '../private-route/private-route-to-favorite';
-import PrivateRouteToLogin from '../private-route/private-route-to-login';
+import PrivateRouteToFavorite from '../private-route-to-favorite/private-route-to-favorite';
+import PrivateRouteToLogin from '../private-route-to-login/private-route-to-login';
 import browserHistory from '../../browser-history';
 import { init } from './actions/init';
 function App(props) {
