+++ project/src/components/card-for-room-component/card-for-room-component.jsx
