+++ project/src/components/card/card.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { ActionCreator } from '../store/action';
+import { ActionCreator } from '../../store/action';
 
 function Card(props) {
