+++ project/src/components/favorite-card/favorite-card.jsx
