+++ project/src/components/favorite-empty/favorites-empty.jsx
