+++ project/src/components/footer-logo/footer-logo.jsx
