+++ project/src/components/header/header.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import NavListLogin from './nav-list-login';
-import NavListForAuth from './nav-list-auth';
-import { AuthorizationStatus } from '../constants';
-import Logo from '../components/logo';
+import NavListLogin from '../nav-list-login/nav-list-login';
+import NavListForAuth from '../nav-list-for-auth/nav-list-for-auth';
+import { AuthorizationStatus } from '../../constants';
+import Logo from '../logo/logo';
 import PropTypes from 'prop-types';
 
 function Header(props) {
