+++ project/src/components/map/map.jsx
@@ -2,14 +2,16 @@ import leaflet from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import React, { useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import useMap from '../useMap';
+import useMap from '../../hooks/use-map';
 import { connect } from 'react-redux';
-import { getCardWithTheSameId } from '../utils';
-import offersInMap from '../prop-types/offers-used-in-map';
+import { getCardWithTheSameId } from '../../utils/utils';
+import offersInMap from '../../prop-types/offers-used-in-map';
 
 const FIRST_ARRAY_ELEMENT = 0;
-const MARKER_URL = 'img/pin.svg';
-const ACTIVE_MARKER_URL = 'img/pin-active.svg';
+const MarkerUrl = {
+  MARKER_URL: 'img/pin.svg',
+  ACTIVE_MARKER_URL: 'img/pin-active.svg',
+};
 
 function Map(props) {
   const { activeCardId, cards, adsList } = props;
@@ -19,13 +21,13 @@ function Map(props) {
   const map = useMap(mapRef, city);
 
   const icon = leaflet.icon({
-    iconUrl: MARKER_URL,
+    iconUrl: MarkerUrl.MARKER_URL,
     iconSize: [30, 30],
     iconAnchor: [15, 30],
   });
 
   const activeIcon = leaflet.icon({
-    iconUrl: ACTIVE_MARKER_URL,
+    iconUrl: MarkerUrl.ACTIVE_MARKER_URL,
     iconSize: [30, 30],
     iconAnchor: [15, 30],
   });
