--- project/src/components/mocks/reviews.js
@@ -1,146 +1,0 @@
-export const Reviews = [
-  {
-    comment: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    date: '2019-07-08T14:13:56.569Z',
-    id: 1,
-    ratingInReview: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      authorName: 'Max',
-    },
-  },
-  {
-    comment: 'Amazing apartment with wonderfull view',
-    date: '2019-05-08T14:13:56.569Z',
-    id: 2,
-    ratingInReview: 3,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      authorName: 'Angi',
-    },
-  },
-  {
-    comment: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    date: '2021-07-08T14:13:56.569Z',
-    id: 12,
-    ratingInReview: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      authorName: 'Max',
-    },
-  },
-  {
-    comment: 'Amazing apartment with wonderfull view',
-    date: '2020-05-08T14:13:56.569Z',
-    id: 13,
-    ratingInReview: 3,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      authorName: 'Angi',
-    },
-  },
-  {
-    comment: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    date: '2019-06-08T14:13:56.569Z',
-    id: 10,
-    ratingInReview: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      authorName: 'Max',
-    },
-  },
-  {
-    comment: 'Amazing apartment with wonderfull view',
-    date: '2017-05-08T14:13:56.569Z',
-    id: 11,
-    ratingInReview: 3,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      authorName: 'Angi',
-    },
-  },
-  {
-    comment: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    date: '2019-07-09T14:13:56.569Z',
-    id: 15,
-    ratingInReview: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      authorName: 'Max',
-    },
-  },
-  {
-    comment: 'Amazing apartment with wonderfull view',
-    date: '2019-11-08T14:13:56.569Z',
-    id: 20,
-    ratingInReview: 3,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      authorName: 'Angi',
-    },
-  },
-  {
-    comment: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    date: '2020-12-08T14:13:56.569Z',
-    id: 21,
-    ratingInReview: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      authorName: 'Max',
-    },
-  },
-  {
-    comment: 'Amazing apartment with wonderfull view',
-    date: '2016-05-19T14:13:56.569Z',
-    id: 22,
-    ratingInReview: 3,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      authorName: 'Angi',
-    },
-  },
-  {
-    comment: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    date: '2019-07-17T14:13:56.569Z',
-    id: 23,
-    ratingInReview: 4,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: false,
-      authorName: 'Max',
-    },
-  },
-  {
-    comment: 'Amazing apartment with wonderfull view',
-    date: '2019-05-16T14:13:56.569Z',
-    id: 24,
-    ratingInReview: 3,
-    user: {
-      avatarUrl: 'img/avatar-angelina.jpg',
-      id: 4,
-      isPro: true,
-      authorName: 'Angi',
-    },
-  },
-];
