+++ project/src/components/new-comment-form/new-comment-form.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { ratingForm } from '../constants';
+import { ratingForm } from '../../constants';
 
 const createRatingForm = (numberValue, definition) => (
   <React.Fragment key={definition}>
