+++ project/src/routes/main-page/selectors.js
@@ -1,0 +1,1 @@
+export const filterCardsByCurrentCity = (cards, city) => cards.filter((card) => city === card.city.name);
