+++ project/src/store/action.js
@@ -13,8 +13,6 @@ export const ActionType = {
   CHANGE_CITY: 'change-city',
   SET_ACTIVE_CARD: 'set-active-card',
   RESET_ACTIVE_CARD: 'reset-active-card',
-  SHOW_SORT_MENU: 'show-sort-menu',
-  RESET_SORT_MENU: 'reset-sort-menu',
   SET_SORT_TYPE: 'set-sort-type',
 };
 
@@ -72,12 +70,6 @@ export const ActionCreator = {
   }),
   resetActiveCard: () => ({
     type: ActionType.RESET_ACTIVE_CARD,
-  }),
-  showSortMenu: () => ({
-    type: ActionType.SHOW_SORT_MENU,
-  }),
-  reserSortMenu: () => ({
-    type: ActionType.RESET_SORT_MENU,
   }),
   setSortType: (sortType) => ({
     type: ActionType.SET_SORT_TYPE,
